+++ lib/services/habit/habit_service.dart
@@ -1,7 +1,9 @@
 import 'package:UHabit/models/habit.dart';
+import 'package:UHabit/models/habit_record.dart';
 
 abstract class HabitService {
   Future<List<Habit>> getPublicHabits();
   Future<Habit?> getHabit(int habitId);
   Future<List<Habit>> getUserHabits(int userId);
+  Future<HabitRecord?> submitHabitRecord(int habitId);
 }
