+++ lib/services/habit/habit_service_impl.dart
@@ -1,33 +1,46 @@
+import 'package:UHabit/models/habit.dart';
 import 'package:UHabit/models/habit_record.dart';
 import 'package:UHabit/services/api_utils.dart';
-import 'package:UHabit/services/habit/habit_record_service.dart';
+import 'habit_service.dart';
 
-class HabitRecordServiceImpl implements HabitRecordService {
-  @override
+class HabitServiceImpl implements HabitService {
+  Future<List<Habit>> getPublicHabits() async {
+    final response = await ApiUtils.get<List<Habit>>(
+      '/api/v1/habits/public',
+      (json) {
+        return (json as List).map((item) => Habit.fromJson(item)).toList();
+      },
+    );
+    return response ?? [];
+  }
+
+  Future<Habit?> getHabit(int habitId) async {
+    return await ApiUtils.get<Habit>(
+      '/api/v1/habits/id/$habitId',
+      (json) => Habit.fromJson(json),
+    );
+  }
+
+  Future<List<Habit>> getUserHabits(int userId) async {
+    final response = await ApiUtils.get<List<Habit>>(
+      '/api/v1/habits/user/$userId',
+      (json) {
+        return (json as List).map((item) => Habit.fromJson(item)).toList();
+      },
+    );
+    return response ?? [];
+  }
+
   Future<List<HabitRecord>> getActiveHabits() async {
     final response = await ApiUtils.get<List<HabitRecord>>(
       '/api/v1/habits/record/active',
-      (json) => (json as List)
-          .map((item) => HabitRecord.fromJson(item))
-          .toList(),
+      (json) {
+        return (json as List).map((item) => HabitRecord.fromJson(item)).toList();
+      },
     );
-    
     return response ?? [];
   }
 
-  @override
-  Future<List<HabitRecord>> getUserHabits(int userId) async {
-    final response = await ApiUtils.get<List<HabitRecord>>(
-      '/api/v1/habits/record/$userId',
-      (json) => (json as List)
-          .map((item) => HabitRecord.fromJson(item))
-          .toList(),
-    );
-    
-    return response ?? [];
-  }
-
-  @override
   Future<HabitRecord?> submitHabitRecord(int habitId) async {
     return await ApiUtils.post<HabitRecord>(
       '/api/v1/habits/record/active/$habitId',
@@ -35,4 +48,4 @@ class HabitRecordServiceImpl implements HabitRecordService {
       (json) => HabitRecord.fromJson(json),
     );
   }
-} 
+}
