+++ lib/services/profile/profile_service.dart
@@ -1,7 +1,7 @@
 import 'package:UHabit/models/user_profile.dart';
 import 'package:UHabit/services/api_utils.dart';
 
-class UserService {
+class ProfileService {
   Future<UserProfile?> getPublicProfile(int userId) async {
     return await ApiUtils.get<UserProfile>(
       '/api/v1/user/public-profile/$userId',
